Move SortOrder to generic resource models
